moved row cell out of table namespace to avoid conflict with costs plugin
