+++ app/cells/views/row/show.erb
